Outputs coverage information in CSV format.
Parsing coverage information may be useful to guarantee that
new versions of packages do not introduce undocumented code.

By default, haddock outputs coverage information using a strange
format like 80% (3 / 4) in 'Some.Module'. This format is not described
anywhere. It would be easier for programs to parse this if it were
printed using CSV-format.

This commit adds a new flag, --csv-coverage, disabled by default, which
makes haddock print those lines using CSV, as follows:
80%, 3, 4, 'Some.Module'.

To be precise, the format for this is:
<percentage%>,<sum_documented>,<sum_total>,'<qualified_module_name>'

Ideally, the percentage should not be included, but the symbol % is
useful to capture these lines using grep. That can be done easily with
the regular expression '^\s*[0-9]\+%', and a further grep -ve '100%'
filters those modules with undocumented entities.

NOTE: Calling haddock with no arguments and `--no-warnings` should only
produce coverage lines (if there are no errors).  Unfortuntely, that may
not be the case when a) invoking haddock using cabal, or b) because
haddock produces some warnings even when `--no-warnings` is used.

In the future, if this changes and a --no-html flag is introduced, I
think the percentage may be removed. It would also be nice to have a
`--coverage-only` flag or an intermediate verbosity level that outputs
this info but not debugging messages. Alternatively, a new flag
`--simple-output` could serve both features (csv and less verbosity).